src/router/requirements.txt
ngram == 3.3.2
protobuf == 3.6.1
protobuf3_to_dict == 0.1.5
pyyaml == 3.13
requests == 2.22.0
requests-aws4auth == 0.9
retrying == 1.3.3
rollbar
six == 1.11.0
unidecode == 1.0.22